Changed existing col distr strategies to other resizing modes.
